+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### v. 1.2.0
+- Changed "QFXparser" class name to "FileParser"
+- Changed "Build" method name to "BuildStatement"
+- Fixed date conversion for different timezones
+
 ### v. 1.1.0
 - Changed "StatementBuilder" class name to "QFXparser"
 - Changed "Transaction" API